Archive containing entire project. See folder "Final Project" for Main code
This is a backup of my working archive of the Game of Life. It was made at near the end of the project - most of the functionality of the project had been implemented. Notably, there is a VGA controller for the Nexsys-2 FPGA provided by Xilinx (it has been slimmed down). The main code of the project is the in the "Final Project" folder. Use ISE Design Suite (V14 was used at the time) to open the project file. Use Adept to program the FPGA using the bitloader function and selecting the corresponding file.